backend/requirements.txt
Django==5.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.0.1
requests==2.31.0
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0
daphne==4.0.0

# AI and Computer Vision packages
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3
dlib==19.24.2